communityapp/requirements.txt
# Frontend
streamlit==1.34.0
PyYAML>=6.0